% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carmen-package.R
\docType{package}
\name{carmen-package}
\alias{carmen}
\alias{carmen-package}
\title{carmen: Where in the World is this IP Address?}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Finds the geographic areas associated with a vector
    of IPv4 addresses.  Uses a local database and vectorized R and C++
    code to provide significant speed imrovements over similar options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rossellhayes/carmen}
  \item Report bugs at \url{https://github.com/rossellhayes/carmen/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
