% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_clear_cache.R
\name{crm_clear_cache}
\alias{crm_clear_cache}
\title{Clear the carmen database cache}
\usage{
crm_clear_cache()
}
\value{
Nothing.
}
\description{
Clear the carmen database cache
}
\seealso{
Other carmen: 
\code{\link{crm_ip_to_int}()},
\code{\link{crm_locate}()}
}
\concept{carmen}
