% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_ip_to_int.R
\name{crm_ip_to_int}
\alias{crm_ip_to_int}
\title{Convert a character vector of IP addresses to integers}
\usage{
crm_ip_to_int(ip)
}
\arguments{
\item{ip}{A character vector of IP addresses (e.g. \code{"130.191.166.222"}).}
}
\value{
A \link{numeric} vector of integer IP addresses.
}
\description{
Note: integer IP addresses are stored as a \link{double} vector to prevent
integer overflows.
}
\examples{
crm_ip_to_int("130.191.166.222")
crm_ip_to_int(generator::r_ipv4_addresses(100))
}
\seealso{
Other carmen: 
\code{\link{crm_locate}()}
}
\concept{carmen}
