% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_locate.R
\name{crm_locate}
\alias{crm_locate}
\title{Where in the world is this IP address?}
\source{
This site or product includes IP2Location LITE data available from
\url{https://lite.ip2location.com}.
}
\usage{
crm_locate(ip, database = ip2location_lite_db1, cache = TRUE)
}
\arguments{
\item{ip}{A vector of IP addresses. Either \link{character}
(e.g. \code{"130.191.166.222"}) or \link{numeric} (e.g. \code{2193598174}).}

\item{database}{A database of IP address locations from
\href{https://ip2location.com/}{IP2Location}.
Either a \link{data.frame} or the path to a \code{.csv} file.}

\item{cache}{If \code{TRUE}, the data in \code{database} will be cached for this
session, improving the speed of subsequent lookups.}
}
\value{
A \link{data.frame} with a row of location information for each element
of \code{ip}.
}
\description{
Where in the world is this IP address?
}
\examples{
crm_locate("130.191.166.222")

library(dplyr)

tibble(
  id   = seq_len(100),
  name = generator::r_full_names(100),
  ip   = generator::r_ipv4_addresses(100)
) \%>\%
  mutate(crm_locate(ip))
}
\seealso{
Other carmen: 
\code{\link{crm_clear_cache}()},
\code{\link{crm_ip_to_int}()}
}
\concept{carmen}
