% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carmen-package.R
\docType{package}
\name{carmen-package}
\alias{carmen}
\alias{carmen-package}
\title{carmen: Where in the World is this IP Address?}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
